// src/app/modules/cnes/interfaces/cnes.interface.ts
export interface CnesInterface {
  bairro_estabelecimento: string;
  codigo_atividade_ensino_unidade: string;
  codigo_cep_estabelecimento: string;
  codigo_cnes: number;
  codigo_esfera_administrativa_unidade?: number;
  codigo_estabelecimento_saude: string;
  codigo_identificador_turno_atendimento: string;
  codigo_motivo_desabilitacao_estabelecimento?: number;
  codigo_municipio: number;
  codigo_natureza_organizacao_unidade?: null;
  codigo_nivel_hierarquia_unidade?: null;
  codigo_tipo_unidade: number;
  codigo_uf: number;
  descricao_esfera_administrativa?: null;
  descricao_natureza_juridica_estabelecimento: string;
  descricao_nivel_hierarquia?: null;
  descricao_turno_atendimento: string;
  endereco_email_estabelecimento: string;
  endereco_estabelecimento: string;
  estabelecimento_faz_atendimento_ambulatorial_sus: string;
  estabelecimento_possui_atendimento_ambulatorial: boolean;
  estabelecimento_possui_atendimento_hospitalar: boolean;
  estabelecimento_possui_centro_cirurgico: boolean;
  estabelecimento_possui_centro_neonatal: boolean;
  estabelecimento_possui_centro_obstetrico: boolean;
  estabelecimento_possui_servico_apoio: boolean;
  latitude_estabelecimento_decimo_grau: number;
  longitude_estabelecimento_decimo_grau: number;
  natureza_organizacao_entidade?: null;
  nome_fantasia: string;
  nome_razao_social: string;
  numero_cnpj?: null;
  numero_cnpj_entidade?: null;
  numero_estabelecimento: string;
  numero_telefone_estabelecimento: string;
  tipo_gestao: string;
}